refactor(watchlist): extract is watchlisted store
