+++ projects/client/src/lib/sections/media-actions/watchlist/useWatchlist.ts
@@ -1,42 +1,29 @@
-import { useUser } from '$lib/features/auth/stores/useUser.ts';
+import type { MediaStoreProps } from '$lib/models/MediaStoreProps.ts';
 import { InvalidateAction } from '$lib/requests/models/InvalidateAction.ts';
-import type { MediaType } from '$lib/requests/models/MediaType.ts';
 import { addToWatchlistRequest } from '$lib/requests/sync/addToWatchlistRequest.ts';
 import { removeFromWatchlistRequest } from '$lib/requests/sync/removeFromWatchlistRequest.ts';
 import { toWatchlistPayload } from '$lib/sections/media-actions/watchlist/toWatchlistPayload.ts';
 import { useInvalidator } from '$lib/stores/useInvalidator.ts';
-import { derived, writable } from 'svelte/store';
+import { writable } from 'svelte/store';
+import { useIsWatchlisted } from './useIsWatchlisted.ts';
 
-export type WatchlistStoreProps = {
-  type: MediaType;
-  media: { id: number };
-};
-
-export function useWatchlist({ type, media }: WatchlistStoreProps) {
+export function useWatchlist(props: MediaStoreProps) {
+  const { type } = props;
+  const media = Array.isArray(props.media) ? props.media : [props.media];
   const isWatchlistUpdating = writable(false);
-  const { watchlist } = useUser();
   const { invalidate } = useInvalidator();
+  const ids = media.map(({ id }) => id);
 
-  const isWatchlisted = derived(
-    watchlist,
-    ($watchlist) => {
-      if (!$watchlist) {
-        return false;
-      }
-
-      switch (type) {
-        case 'movie':
-          return $watchlist.movies.has(media.id);
-        case 'show':
-          return $watchlist.shows.has(media.id);
-      }
-    },
-  );
+  const { isWatchlisted } = useIsWatchlisted(props);
 
   const addToWatchlist = async () => {
+    if (type === 'episode') {
+      return;
+    }
+
     isWatchlistUpdating.set(true);
     await addToWatchlistRequest({
-      body: toWatchlistPayload(type, [media.id]),
+      body: toWatchlistPayload(type, ids),
     });
 
     await invalidate(InvalidateAction.Watchlisted(type));
@@ -45,9 +32,13 @@ export function useWatchlist({ type, media }: WatchlistStoreProps) {
   };
 
   const removeFromWatchlist = async () => {
+    if (type === 'episode') {
+      return;
+    }
+
     isWatchlistUpdating.set(true);
     await removeFromWatchlistRequest({
-      body: toWatchlistPayload(type, [media.id]),
+      body: toWatchlistPayload(type, ids),
     });
 
     await invalidate(InvalidateAction.Watchlisted(type));
